Use MIT license
We use the MIT license here because we want to be permissive with what
other can do with the work in the repository. The goal of this repo is
to help new and experienced users find better ways to make git a useful
tool for communication, collaboration, and code organization. Being
restrictive with the content seems antithetical to its goals so we
therefore use the MIT license.